Add prettier to project
To improve our code formatting this change adds a new format command
that will run our prettier options against the codebase. This is
supposed to act as a complement to eslint and fix simple errors for us.
In addition to adding the command this will also format your staged
changes when you commit anything new.